Smart glove
The system is based on sensor gloves which convert the sign language to speech. The glove is 
equipped with flex sensors and an accelerometer which senses the gesture. The controller identify the 
gesture and transmits a character corresponding to this gesture to an android device. And the respective 
voice for the character is speeches out through the android device.